+++ sia/controllers/lqr.cpp
@@ -29,37 +29,38 @@ const Eigen::VectorXd& LQR::policy(const Distribution& state) {
   m_states.clear();
   m_states.reserve(T);
 
-  // 1. Backward recursion dynamic Ricatti equation to compute the cost to go
-  std::vector<Eigen::MatrixXd> feedback;
-  std::vector<Eigen::VectorXd> feedforward;
-  feedback.reserve(T);
-  feedforward.reserve(T);
+  m_feedforward.clear();
+  m_feedforward.reserve(T);
+  m_feedback.clear();
+  m_feedback.reserve(T);
 
+  // 1. Backward recursion dynamic Ricatti equation to compute the cost to go
   // Initialize terminal value function Gradient and Hessian
   Eigen::MatrixXd P = Qf;
-  Eigen::VectorXd v = Qf * m_cost.xd(T - 1);
+  Eigen::VectorXd v = -Qf * m_cost.xd(T);
   Eigen::MatrixXd QuuInv;
 
-  for (int i = T - 1; i >= 0; --i) {
-    bool r = svdInverse(R + G.transpose() * P * G, QuuInv);
+  for (int k = T - 1; k >= 0; --k) {
+    const Eigen::MatrixXd Quu = R + G.transpose() * P * G;
+    bool r = svdInverse(Quu, QuuInv);
     SIA_EXCEPTION(r, "Matrix inversion failed in LQR cost to go computation");
-    const Eigen::MatrixXd K = QuuInv * G.transpose() * P * F;
-    const Eigen::VectorXd k = QuuInv * G.transpose() * v;
-    feedback.emplace_back(K);
-    feedforward.emplace_back(k);
+    const Eigen::MatrixXd K = -QuuInv * G.transpose() * P * F;
+    const Eigen::VectorXd d = -QuuInv * G.transpose() * v;
+    m_feedback.emplace_back(K);
+    m_feedforward.emplace_back(d);
 
-    const Eigen::MatrixXd FGK = F - G * K;
+    const Eigen::MatrixXd FGK = F + G * K;
     P = F.transpose() * P * FGK + Q;
-    v = FGK.transpose() * v + Qf * m_cost.xd(i);
+    v = FGK.transpose() * v - Q * m_cost.xd(k);
   }
 
   // Flip the gain vector/matrices order to advance forward in time
-  std::reverse(feedback.begin(), feedback.end());
-  std::reverse(feedforward.begin(), feedforward.end());
+  std::reverse(m_feedback.begin(), m_feedback.end());
+  std::reverse(m_feedforward.begin(), m_feedforward.end());
 
   // 2. Forward pass to compute the optimal control and states
-  for (std::size_t i = 0; i < T; ++i) {
-    const Eigen::VectorXd u = -feedback.at(i) * x + feedforward.at(i);
+  for (std::size_t k = 0; k < T; ++k) {
+    const Eigen::VectorXd u = m_feedback.at(k) * x + m_feedforward.at(k);
     m_controls.emplace_back(u);
     m_states.emplace_back(x);
     x = m_dynamics.f(x, u);
@@ -74,6 +75,14 @@ const Trajectory<Eigen::VectorXd>& LQR::controls() const {
 
 const Trajectory<Eigen::VectorXd>& LQR::states() const {
   return m_states;
+}
+
+const Trajectory<Eigen::VectorXd>& LQR::feedforward() const {
+  return m_feedforward;
+}
+
+const Trajectory<Eigen::MatrixXd>& LQR::feedback() const {
+  return m_feedback;
 }
 
 }  // namespace sia